backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
pillow>=10.0.0
deepface
opencv-python-headless>=4.8.0
numpy>=1.24.0
tf-keras
